UsersAPI/requirements.txt
Django==1.11
djangorestframework==3.8.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.4
Pillow==5.1.0

